requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
numpy>=1.23.0
xgboost>=1.7.0
matplotlib>=3.7.0
shap>=0.44.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
Pillow>=9.3.0  # For image file support in file uploader
openai
langchain
langchain-community
sentence-transformers
faiss-cpu
PyMuPDF


